Refactor Diff/Diff3 Slightly
