+++ SemDiff.Core/SemDiff.Core.csproj
@@ -115,6 +115,7 @@
     <Compile Include="ChangedFile.cs" />
     <Compile Include="Conflict.cs" />
     <Compile Include="ConflictInfo.cs" />
+    <Compile Include="Diff.cs" />
     <Compile Include="Diff3.cs" />
     <Compile Include="GitHub.cs" />
     <Compile Include="Logger.cs" />
